requirements.txt
flask
flask-cors
python-dotenv
supabase
python-dotenv